make new Home page original
