+++ app/src/screens/HomeScreen/index0.js
@@ -1,12 +1,5 @@
 import React, { Component } from "react";
-import {
-  Text,
-  View,
-  StyleSheet,
-  Image,
-  ScrollView,
-  SafeAreaView,
-} from "react-native";
+import { Text, View, StyleSheet, Alert } from "react-native";
 import { Colors, TouchableRipple } from "react-native-paper";
 import { connect } from "react-redux";
 import Ionicons from "@expo/vector-icons/Ionicons";
@@ -24,7 +17,7 @@ import { TouchableWithoutFeedback } from "react-native-gesture-handler";
 
 import { jorneyAction } from "../../store/actions/journey_actions";
 
-export class Test extends Component {
+export class HomeScreen extends Component {
   state = {
     text: "",
     showModal: false,
@@ -113,11 +106,7 @@ export class Test extends Component {
     let toInputTextColor = _.isEmpty(this.state.To) ? "black" : "white";
 
     return (
-      <ScrollView
-        contentContainerStyle={{ flexGrow: 1 }}
-        // stickyHeaderIndices={[2]}
-        showsVerticalScrollIndicator={false}
-      >
+      <>
         <AddressModal
           handleModalVisibility={this.handleModalVisibility}
           showModal={this.state.showModal}
@@ -129,147 +118,132 @@ export class Test extends Component {
           handleCustomPicker={this.handleCustomPicker}
           pickedDateForPicker={this.state.pickedDateForPicker}
         />
-        <SafeAreaView
-          style={{
-            height: 250,
-            backgroundColor: ConstantColors.tintColor,
-            justifyContent: "center",
-          }}
-        >
+        <HomeHeader headerTitle="" />
+        <View style={{ flex: 1 }}>
           <Banner />
-        </SafeAreaView>
-        <View style={{ height: 380, alignItems: "center", marginTop: -100 }}>
-          <Card style={{ height: "100%", width: "90%" }}>
-            {/* <View style={styles.searchbarContainer}></View> */}
-            <View style={{ height: "50%", marginHorizontal: 20 }}>
-              <TouchableRipple
-                style={{
-                  ...styles.inputField,
-                  backgroundColor: fromInputFieldBackground,
-                }}
-                onPress={() =>
-                  this.handleModalVisibility({ addressState: "From" })
-                }
-              >
-                <Text
-                  style={{ ...styles.inputText, color: fromInputTextColor }}
+          <View style={styles.container}>
+            <Card style={{ height: "55%", width: "90%", marginTop: 10 }}>
+              {/* <View style={styles.searchbarContainer}></View> */}
+              <View style={{ height: "50%", marginHorizontal: 20 }}>
+                <TouchableRipple
+                  style={{
+                    ...styles.inputField,
+                    backgroundColor: fromInputFieldBackground,
+                  }}
+                  onPress={() =>
+                    this.handleModalVisibility({ addressState: "From" })
+                  }
                 >
-                  {this.state.From.name || "FROM"}
-                </Text>
-              </TouchableRipple>
-              <TouchableRipple
-                style={{
-                  ...styles.inputField,
-                  backgroundColor: toInputFieldBackground,
-                }}
-                onPress={() =>
-                  this.handleModalVisibility({ addressState: "To" })
-                }
-              >
-                <Text style={{ ...styles.inputText, color: toInputTextColor }}>
-                  {this.state.To.name || "TO"}
-                </Text>
-              </TouchableRipple>
+                  <Text
+                    style={{ ...styles.inputText, color: fromInputTextColor }}
+                  >
+                    {this.state.From.name || "FROM"}
+                  </Text>
+                </TouchableRipple>
+                <TouchableRipple
+                  style={{
+                    ...styles.inputField,
+                    backgroundColor: toInputFieldBackground,
+                  }}
+                  onPress={() =>
+                    this.handleModalVisibility({ addressState: "To" })
+                  }
+                >
+                  <Text
+                    style={{ ...styles.inputText, color: toInputTextColor }}
+                  >
+                    {this.state.To.name || "TO"}
+                  </Text>
+                </TouchableRipple>
 
-              <View style={styles.pickerContainer}>
-                <Text style={{ fontSize: 20 }}>
-                  {this.state.pickedDate || "Pick a date!"}
-                </Text>
-                <View style={styles.datePickerCircleContainer}>
-                  <TouchableWithoutFeedback
-                    style={{
-                      ...styles.circleButton,
-                      backgroundColor:
-                        this.state.pickedColor === "today"
-                          ? ConstantColors.bannerColor
-                          : ConstantColors.initialColor,
-                    }}
-                    onPress={() => this.handleDatePicker("today")}
-                  >
-                    <Text
+                <View style={styles.pickerContainer}>
+                  <Text style={{ fontSize: 20 }}>
+                    {this.state.pickedDate || "Pick a date!"}
+                  </Text>
+                  <View style={styles.datePickerCircleContainer}>
+                    <TouchableWithoutFeedback
                       style={{
-                        color:
+                        ...styles.circleButton,
+                        backgroundColor:
                           this.state.pickedColor === "today"
-                            ? "white"
-                            : ConstantColors.tintColor,
+                            ? ConstantColors.bannerColor
+                            : ConstantColors.initialColor,
                       }}
+                      onPress={() => this.handleDatePicker("today")}
                     >
-                      Today
-                    </Text>
-                  </TouchableWithoutFeedback>
-                  <TouchableWithoutFeedback
-                    style={{
-                      ...styles.circleButton,
-                      backgroundColor:
-                        this.state.pickedColor === "tomorrow"
-                          ? ConstantColors.bannerColor
-                          : ConstantColors.initialColor,
-                    }}
-                    onPress={() => this.handleDatePicker("tomorrow")}
-                  >
-                    <Text
+                      <Text
+                        style={{
+                          color:
+                            this.state.pickedColor === "today"
+                              ? "white"
+                              : ConstantColors.tintColor,
+                        }}
+                      >
+                        Today
+                      </Text>
+                    </TouchableWithoutFeedback>
+                    <TouchableWithoutFeedback
                       style={{
-                        color:
+                        ...styles.circleButton,
+                        backgroundColor:
                           this.state.pickedColor === "tomorrow"
-                            ? "white"
-                            : ConstantColors.tintColor,
+                            ? ConstantColors.bannerColor
+                            : ConstantColors.initialColor,
                       }}
+                      onPress={() => this.handleDatePicker("tomorrow")}
                     >
-                      Tomorrow
-                    </Text>
-                  </TouchableWithoutFeedback>
-                  <TouchableWithoutFeedback
-                    style={{
-                      ...styles.circleButton,
-                      backgroundColor:
-                        this.state.pickedColor === "custom"
-                          ? ConstantColors.bannerColor
-                          : ConstantColors.initialColor,
+                      <Text
+                        style={{
+                          color:
+                            this.state.pickedColor === "tomorrow"
+                              ? "white"
+                              : ConstantColors.tintColor,
+                        }}
+                      >
+                        Tomorrow
+                      </Text>
+                    </TouchableWithoutFeedback>
+                    <TouchableWithoutFeedback
+                      style={{
+                        ...styles.circleButton,
+                        backgroundColor:
+                          this.state.pickedColor === "custom"
+                            ? ConstantColors.bannerColor
+                            : ConstantColors.initialColor,
+                      }}
+                      onPress={() => this.handleDatePicker("custom")}
+                    >
+                      <Ionicons
+                        name="ios-calendar"
+                        size={20}
+                        color={
+                          this.state.pickedColor === "custom"
+                            ? "white"
+                            : ConstantColors.tintColor
+                        }
+                      />
+                    </TouchableWithoutFeedback>
+                  </View>
+                </View>
+                <View style={styles.submitButtonContainer}>
+                  <Button
+                    mode="contained"
+                    onPress={() => {
+                      this.props.jorneyAction(this.state);
+                      this.props.navigation.navigate("Search");
                     }}
-                    onPress={() => this.handleDatePicker("custom")}
+                    style={styles.submitButton}
+                    disabled={!buttonEnabled}
                   >
-                    <Ionicons
-                      name="ios-calendar"
-                      size={20}
-                      color={
-                        this.state.pickedColor === "custom"
-                          ? "white"
-                          : ConstantColors.tintColor
-                      }
-                    />
-                  </TouchableWithoutFeedback>
+                    <Text style={{ fontSize: 20 }}>GO !</Text>
+                  </Button>
                 </View>
               </View>
-              <View style={styles.submitButtonContainer}>
-                <Button
-                  mode="contained"
-                  onPress={() => {
-                    this.props.jorneyAction(this.state);
-                    this.props.navigation.navigate("Search");
-                  }}
-                  style={styles.submitButton}
-                  disabled={!buttonEnabled}
-                >
-                  <Text style={{ fontSize: 20 }}>GO !</Text>
-                </Button>
-              </View>
-            </View>
-          </Card>
-        </View>
-        <View style={{ height: 350 }}>
-          <View style={{ flex: 0.05 }}></View>
-          <View style={{ flex: 0.3 }}>
-            <Image
-              source={require("../../../assets/busbookingbanner.png")}
-              style={{ height: "100%" }}
-            ></Image>
-          </View>
-          <View style={{ flex: 0.05 }}></View>
-          <View style={{ flex: .6 }}>
+            </Card>
             <AllBusScrollView />
           </View>
         </View>
-      </ScrollView>
+      </>
     );
   }
 }
@@ -331,7 +305,7 @@ const styles = StyleSheet.create({
   },
 
   submitButtonContainer: {
-    marginTop: 35,
+    marginTop: 25,
     height: 40,
     justifyContent: "center",
     alignItems: "center",
@@ -358,4 +332,4 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators({ jorneyAction }, dispatch);
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Test);
+export default connect(mapStateToProps, mapDispatchToProps)(HomeScreen);
